Added a badge of the RubyNLP network.
Dear @brenes,
I'd like to include your project into the RubyNLP awesome list. One possible backtrace is the RubyNLP badge in the title of your project. Let's grow the NLP community for Ruby together!
Thank you for your work and consider merging this PR. Feel free to not do so if you think this badge is obtrusive or unsuitable for your README file.